bicep parser and converter new updates
