+++ pkg/parser/bicep/parser.go
@@ -3,9 +3,6 @@ package bicep
 import (
 	"encoding/json"
 	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
 
 	"github.com/Checkmarx/kics/pkg/model"
 )
@@ -14,26 +11,88 @@ import (
 type Parser struct {
 }
 
-// Parse - parses bicep to ARM template (json file)
-func (p *Parser) Parse(bicepPath string, _ []byte) ([]model.Document, []int, error) {
+type ElemBicep struct {
+	Type     string
+	Param    Param
+	Metadata *Metadata
+	Variable Variable
+	Resource Resource
+	Output   Output
+	Module   Module
+}
+
+type Metadata struct {
+	Description string `json:"description"`
+	Name        string `json:"name"`
+}
+
+type Param struct {
+	Name         string    `json:"paramName"`
+	Type         string    `json:"paramType"`
+	DefaultValue string    `json:"paramDefaultValue"`
+	Metadata     *Metadata `json:"paramMetadata"`
+}
+
+type Variable struct {
+	Name        string `json:"varName"`
+	Type        string `json:"varType"`
+	Description string `json:"varDescription"`
+}
+
+type Resource struct {
+	Name        string `json:"resName"`
+	Type        string `json:"resType"`
+	Description string `json:"resDescription"`
+}
+
+type Output struct {
+	Name        string `json:"outName"`
+	Type        string `json:"outType"`
+	Description string `json:"outDescription"`
+}
+
+type Module struct {
+	Name        string `json:"modName"`
+	Path        string `json:"modPath"`
+	Description string `json:"modDescription"`
+}
+
+// Parse - parses bicep to JSON_Bicep template (json file)
+func (p *Parser) Parse(bicepPath string, fileContent []byte) ([]model.Document, []int, error) {
 
 	doc := model.Document{}
+	//reader := bytes.NewReader(fileContent)
+	//scanner := bufio.NewScanner(reader)
+	//elems, err := parserBicepFile(bicepPath)
+	elems := []ElemBicep{
+		{
+			Resource: Resource{
+				Name:        "test_resource",
+				Type:        "Microsoft.Web/sites",
+				Description: "Description",
+			},
+			Type:     "resource",
+			Param:    Param{},
+			Metadata: &Metadata{},
+			Variable: Variable{},
+			Output:   Output{},
+			Module:   Module{},
+		},
+	}
 
-	fileContent, err := convertBicepToArm(bicepPath)
+	jBicep, err := Convert(elems)
+	if err != nil {
+		return nil, nil, err
+	}
+	fmt.Println("jBicep:", jBicep)
+
+	bicepBytes, err := json.Marshal(jBicep)
 
 	if err != nil {
 		return nil, nil, err
 	}
 
-	BicepSyntax, err := teste(bicepPath)
-	if err != nil {
-		return nil, nil, err
-	}
-	fmt.Println("Parser function bicepSyntax:", BicepSyntax)
-
-	err = json.Unmarshal(fileContent, &doc)
-
-	fmt.Println("Parser function doc:", doc)
+	err = json.Unmarshal(bicepBytes, &doc)
 
 	if err != nil {
 		return nil, nil, err
@@ -43,6 +102,63 @@ func (p *Parser) Parse(bicepPath string, _ []byte) ([]model.Document, []int, err
 
 }
 
+/*
+func parseResource(line string) *Resource {
+
+	resourceRegex := regexp.MustCompile(`^resource\s+(\S+)\s+'(\S+)'`)
+	matches := resourceRegex.FindStringSubmatch(line)
+	if matches != nil {
+		resourceType := strings.Split(matches[2], "@")[0]
+		resourceType = strings.ReplaceAll(resourceType, "'", "")
+		return &Resource{
+			Name:        matches[1],
+			Type:        resourceType,
+			Description: "Description",
+		}
+	}
+
+	return nil
+}
+*/
+
+/*
+func parserBicepFile(bicepFile string) ([]ElemBicep, error) {
+
+	file, err := os.Open(bicepFile)
+	if err != nil {
+		return []ElemBicep{}, err
+	}
+
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+	//modules := []Module{}
+	resources := []Resource{}
+
+	for scanner.Scan() {
+		line := scanner.Text()
+
+		if line == "" {
+			continue
+		}
+
+		resource := parseResource(line)
+		if resource != nil {
+			resources = append(resources, *resource)
+			continue
+		}
+
+	}
+
+	if err := scanner.Err(); err != nil {
+		return []ElemBicep{}, err
+	}
+
+	return []ElemBicep{}, nil
+}
+*/
+
+/*
 func convertBicepToArm(bicepPath string) ([]byte, error) {
 
 	armPath := filepath.Join("./teste", "arm.json")
@@ -64,6 +180,7 @@ func convertBicepToArm(bicepPath string) ([]byte, error) {
 
 	return fileContent, nil
 }
+*/
 
 // GetKind returns the kind of the parser
 func (p *Parser) GetKind() model.FileKind {
